Revert ":recycle: move packages"
This reverts commit 5cfb56130d69841dc472aa1494e343860e6a3621.
