+++ app/handler/handler_interface.go
@@ -3,9 +3,9 @@ package handler
 import (
 	"context"
 
-	"github.com/takahiroaoki/grpc-sample/app/domain/repository"
-	"github.com/takahiroaoki/grpc-sample/app/domain/service"
 	"github.com/takahiroaoki/grpc-sample/app/pb"
+	"github.com/takahiroaoki/grpc-sample/app/repository"
+	"github.com/takahiroaoki/grpc-sample/app/service"
 )
 
 type Handler[Req, Res any] interface {
