+++ app/service/get_user_info_service.go
@@ -1,8 +1,8 @@
 package service
 
 import (
-	"github.com/takahiroaoki/grpc-sample/app/domain/entity"
-	"github.com/takahiroaoki/grpc-sample/app/domain/repository"
+	"github.com/takahiroaoki/grpc-sample/app/entity"
+	"github.com/takahiroaoki/grpc-sample/app/repository"
 )
 
 type getUserInfoServiceImpl struct{}
